Fix for pstaff reviving pets and more
There was no check for HP for both of the sides `attacking` and `ally`. This caused buffs being removed from dead enemy and dealing damage to them, debuffs being removed from dead ally and that caused reviving the pet.

Not sure if there are more changes needed, such as selecting different ally or enemy.